requirements.txt
torch
torchvision
torchaudio
numpy
opencv-python
scikit-image
pillow
psutil
keyboard